Made the windows port not delete the channel guide html file.  This fixes the back button more or less.
Removed the hardcoded channel guide URL and replaced it with a config.get call
in urlcallbacks.



git-svn-id: https://svn.participatoryculture.org/svn/dtv/branches/Democracy-Player-0.8.5@2326 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: 66be637d5c3d48fe20f468e2b30e0185a2f79a7f

+++ tv/platform/windows-xul/frontend_implementation/HTMLDisplay.py
@@ -7,9 +7,10 @@ import errno
 import os
 
 import app
+import config
 import tempfile
 import os
-
+import prefs
 import frontend
 from frontend_implementation import urlcallbacks
 
@@ -58,6 +59,11 @@ class HTMLDisplay (app.Display):
         self.pageLoadFinised = False
         self.deferedCalls = []
         self.location = None
+        if baseURL == config.get(prefs.CHANNEL_GUIDE_URL):
+            self.removeTempFile = False
+        else:
+            self.removeTempFile = True
+
 
     def setInitialHTML(self):
         if not os.path.exists(tempdir):
@@ -81,10 +87,11 @@ class HTMLDisplay (app.Display):
             self.pageLoadFinised = True
             for function, args in self.deferedCalls:
                 function(self, *args)
-            try:
-                os.unlink(self.location)
-            except:
-                pass
+            if self.removeTempFile:
+                try:
+                    os.unlink(self.location)
+                except:
+                    pass
 
     def setArea(self, area):
         self.area = area
